VueApi/requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
et-xmlfile==1.1.0
greenlet==1.1.1
importlib-metadata==4.8.1
iniconfig==1.1.1
myconfig==0.2.7
numpy==1.21.2
openpyxl==3.0.7
packaging==21.0
pandas==1.3.2
pluggy==1.0.0
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
selenium==3.141.0
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
xlrd==1.2.0
zipp==3.5.0
